Rewamp history screen for desktop
