+++ lib/l10n/app_en.arb
@@ -47,7 +47,7 @@
   "selectACamera": "Select Camera",
   "online": "Online",
   "offline": "Offline",
-  "eventBrowser": "History",
+  "eventBrowser": "Events History",
   "lastUpdate": "Last Update",
   "theme": "Theme",
   "system": "System",
